user destroy route done, did event controlle set up, and create route done (might need some tweaking), can only access if logged in

+++ controllers/userController.js
@@ -61,16 +61,17 @@ router.put('/:id', async (req, res, next) => {
 })
 
 // destroy -- destroy all events and bands associated 
+// right now it just delets the user it self - have to add delete the others later
 router.delete('/:id', async (req, res, next) => {
   try {
-  		// if(req.session.userId == req.params.id){
+  		if(req.session.userId == req.params.id){
   			// console.log(user)
   			await User.findByIdAndRemove(req.params.id)
   			console.log(req.session)
   			res.redirect('/auth/logout')
-  		// }else{
-  		// 	res.direct('/auth/login')
-  		// }
+  		}else{
+  			res.direct('/auth/login')
+  		}
   	}catch(error){
   		next(error)
   	}
